Adopted @tristanpemble Suggested Fix
Due to the agreement to use llvm as a fix for the linux and potentially for the Mac,
I have replaced the dynamic linkage with using llvm instead. On Windows, it compile fine,
passed the test, and it works without any issues on my toy project that uses zgpu;
however, futher testing is required to ensure linux and mac version of the library in
working state.